Update docs for new CLI
